% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicateContentR.R
\name{duplicate_content_score}
\alias{duplicate_content_score}
\title{Get Duplicate Content Ratio between 2 web pages}
\usage{
duplicate_content_score(url_a, url_b, my_user_agent)
}
\arguments{
\item{url_a}{url website A}

\item{url_b}{url website B}

\item{my_user_agent}{user agent}
}
\value{
score
}
\description{
duplicateContentR takes 2 urls as input and computes a duplicate content score to detect plagiarism.
}
\examples{
duplicate_content_score("https://fr.wikipedia.org/wiki/Physique_quantique", "https://www.futura-sciences.com/sciences/definitions/physique-physique-quantique-13197/","Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:61.0) Gecko/20100101 Firefox/61.0")

}
